wip: factor out selectors from labware-ingred
