+++ protocol-designer/src/components/LiquidPlacementModal.js
@@ -13,7 +13,7 @@ import LiquidPlacementForm from '../components/LiquidPlacementForm'
 import SingleLabwareWrapper from '../components/SingleLabware'
 import WellSelectionInstructions from './WellSelectionInstructions'
 
-import {selectors} from '../labware-ingred/reducers'
+import {selectors} from '../labware-ingred/selectors'
 import * as wellContentsSelectors from '../top-selectors/well-contents'
 import wellSelectionSelectors from '../well-selection/selectors'
 import {
@@ -75,7 +75,7 @@ class LiquidPlacementModal extends React.Component<Props, State> {
 const mapStateToProps = (state: BaseState): SP => {
   const containerId = selectors.getSelectedLabwareId(state)
   const selectedWells = wellSelectionSelectors.getSelectedWells(state)
-  if (containerId === null) {
+  if (containerId == null) {
     console.error('LiquidPlacementModal: No labware is selected, and no labwareId was given to LiquidPlacementModal')
     return {
       selectedWells: {},
