+++ protocol-designer/src/components/StepEditForm/TipPositionInput/index.js
@@ -3,7 +3,7 @@ import * as React from 'react'
 import {connect} from 'react-redux'
 import {HoverTooltip, FormGroup, InputField} from '@opentrons/components'
 import { getLabware } from '@opentrons/shared-data'
-import {selectors as labwareIngredsSelectors} from '../../../labware-ingred/reducers'
+import {selectors as labwareIngredsSelectors} from '../../../labware-ingred/selectors'
 import i18n from '../../../localization'
 import {selectors as stepFormSelectors} from '../../../step-forms'
 import stepFormStyles from '../StepEditForm.css'
