+++ protocol-designer/src/containers/SelectorDebugger.js
@@ -4,7 +4,7 @@ import * as React from 'react'
 import { connect } from 'react-redux'
 import mapValues from 'lodash/mapValues'
 
-import {selectors as labwareIngred} from '../labware-ingred/reducers'
+import {selectors as labwareIngred} from '../labware-ingred/selectors'
 import {selectors as fileDataSelectors} from '../file-data'
 import {selectors as stepSelectors} from '../ui/steps'
 import type {BaseState} from '../types'
