+++ protocol-designer/src/labware-ingred/__tests__/selectors.test.js
@@ -1,4 +1,5 @@
-import {selectors} from '../reducers'
+// @flow
+import {selectors} from '../selectors'
 
 // FIXTURES
