+++ protocol-designer/src/labware-ingred/reducers/index.js
@@ -1,35 +1,18 @@
 // @flow
 import {combineReducers} from 'redux'
 import {handleActions, type ActionType} from 'redux-actions'
-import {createSelector} from 'reselect'
-
-import forEach from 'lodash/forEach'
 import omit from 'lodash/omit'
 import mapValues from 'lodash/mapValues'
-import max from 'lodash/max'
 import pickBy from 'lodash/pickBy'
 import reduce from 'lodash/reduce'
 
-import {sortedSlotnames, FIXED_TRASH_ID} from '../../constants.js'
-import {labwareToDisplayName} from '../utils'
-
-import type {DeckSlot} from '@opentrons/components'
-import {getIsTiprack} from '@opentrons/shared-data'
-
-import type {SingleLabwareLiquidState, LabwareLiquidState} from '../../step-generation'
-
-import type {
-  IngredInputs,
-  LiquidGroupsById,
-  AllIngredGroupFields,
-  LiquidGroup,
-  OrderedLiquids,
-  Labware,
-  LabwareTypeById,
-} from '../types'
 import * as actions from '../actions'
+import {_loadedContainersBySlot} from '../utils'
+import {sortedSlotnames, FIXED_TRASH_ID} from '../../constants'
 import {getPDMetadata} from '../../file-types'
-import type {BaseState, Options} from '../../types'
+import type {DeckSlot} from '@opentrons/components'
+import type {SingleLabwareLiquidState, LabwareLiquidState} from '../../step-generation'
+import type {LiquidGroupsById, Labware} from '../types'
 import type {LoadFileAction} from '../../load-file'
 import type {
   RemoveWellsContents,
@@ -63,7 +46,7 @@ const modeLabwareSelection = handleActions({
   CREATE_CONTAINER: () => false,
 }, false)
 
-type SelectedContainerId = string | null
+export type SelectedContainerId = ?string
 const selectedContainerId = handleActions({
   OPEN_INGREDIENT_SELECTOR: (state, action: ActionType<typeof actions.openIngredientSelector>): SelectedContainerId => action.payload,
   CLOSE_INGREDIENT_SELECTOR: (state, action: ActionType<typeof actions.closeIngredientSelector>): SelectedContainerId => null,
@@ -74,13 +57,13 @@ const selectedContainerId = handleActions({
   CLOSE_WELL_SELECTION_MODAL: (): SelectedContainerId => null,
 }, null)
 
-type DrillDownLabwareId = string | null
+export type DrillDownLabwareId = ?string
 const drillDownLabwareId = handleActions({
   DRILL_DOWN_ON_LABWARE: (state, action: ActionType<typeof actions.drillDownOnLabware>): DrillDownLabwareId => action.payload,
   DRILL_UP_FROM_LABWARE: (state, action: ActionType<typeof actions.drillUpFromLabware>): DrillDownLabwareId => null,
 }, null)
 
-type ContainersState = {
+export type ContainersState = {
   [id: string]: ?Labware,
 }
 
@@ -316,229 +299,5 @@ const rootReducer = combineReducers({
   ingredients,
   ingredLocations,
 })
-
-type RootSlice = {labwareIngred: RootState}
-type Selector<T> = (RootSlice) => T
-// SELECTORS
-const rootSelector = (state: RootSlice): RootState => state.labwareIngred
-
-const getLabwareById: Selector<{[labwareId: string]: ?Labware}> = createSelector(
-  rootSelector,
-  rootState => rootState.containers
-)
-
-const getLabwareNicknamesById: Selector<{[labwareId: string]: string}> = createSelector(
-  getLabwareById,
-  (labwareById) => mapValues(
-    labwareById,
-    labwareToDisplayName,
-  )
-)
-
-const getLabwareTypes: Selector<LabwareTypeById> = createSelector(
-  getLabwareById,
-  (labwareById) => mapValues(
-    labwareById,
-    (labware: Labware) => labware.type
-  )
-)
-
-const getLiquidGroupsById = (state: RootSlice) => rootSelector(state).ingredients
-const getLiquidsByLabwareId = (state: RootSlice) => rootSelector(state).ingredLocations
-
-const getNextLiquidGroupId: Selector<string> = createSelector(
-  getLiquidGroupsById,
-  (ingredGroups) => ((max(Object.keys(ingredGroups).map(id => parseInt(id))) + 1) || 0).toString()
-)
-
-const getLiquidNamesById: Selector<{[ingredId: string]: string}> = createSelector(
-  getLiquidGroupsById,
-  ingredGroups => mapValues(ingredGroups, (ingred: LiquidGroup) => ingred.name)
-)
-
-const getLiquidSelectionOptions: Selector<Options> = createSelector(
-  getLiquidGroupsById,
-  (liquidGroupsById) => {
-    return Object.keys(liquidGroupsById).map(id => ({
-      // NOTE: if these fallbacks are used, it's a bug
-      name: (liquidGroupsById[id])
-        ? liquidGroupsById[id].name || `(Unnamed Liquid: ${String(id)})`
-        : 'Missing Liquid',
-      value: id,
-    }))
-  }
-)
-
-const _loadedContainersBySlot = (containers: ContainersState) =>
-  reduce(containers, (acc, labware: ?Labware) => (labware && labware.slot)
-    ? {...acc, [labware.slot]: labware.type}
-    : acc
-    , {})
-
-const loadedContainersBySlot = createSelector(
-  getLabwareById,
-  containers => _loadedContainersBySlot(containers)
-)
-
-/** Returns options for dropdowns, excluding tiprack labware */
-const labwareOptions: Selector<Options> = createSelector(
-  getLabwareById,
-  getLabwareNicknamesById,
-  (labwareById, names) => reduce(labwareById, (acc: Options, labware: Labware, labwareId): Options => {
-    const isTiprack = getIsTiprack(labware.type)
-    if (!labware.type || isTiprack) {
-      return acc
-    }
-    return [
-      ...acc,
-      {
-        name: names[labwareId],
-        value: labwareId,
-      },
-    ]
-  }, [])
-)
-
-const DISPOSAL_LABWARE_TYPES = ['trash-box', 'fixed-trash']
-/** Returns options for disposal (e.g. fixed trash and trash box) */
-const disposalLabwareOptions: Selector<Options> = createSelector(
-  getLabwareById,
-  getLabwareNicknamesById,
-  (labwareById, names) => reduce(labwareById, (acc: Options, labware: Labware, labwareId): Options => {
-    if (!labware.type || !DISPOSAL_LABWARE_TYPES.includes(labware.type)) {
-      return acc
-    }
-    return [
-      ...acc,
-      {
-        name: names[labwareId],
-        value: labwareId,
-      },
-    ]
-  }, [])
-)
-
-// false or selected slot to add labware to, eg 'A2'
-const selectedAddLabwareSlot = (state: BaseState) => rootSelector(state).modeLabwareSelection
-
-const getSavedLabware = (state: BaseState) => rootSelector(state).savedLabware
-
-const getSelectedLabwareId: Selector<SelectedContainerId> = createSelector(
-  rootSelector,
-  rootState => rootState.selectedContainerId
-)
-
-const getSelectedLiquidGroupState: Selector<SelectedLiquidGroupState> = createSelector(
-  rootSelector,
-  rootState => rootState.selectedLiquidGroup
-)
-
-const getSelectedLabware: Selector<?Labware> = createSelector(
-  getSelectedLabwareId,
-  getLabwareById,
-  (selectedLabwareId, labware) =>
-    (selectedLabwareId && labware[selectedLabwareId]) || null
-)
-
-const getDrillDownLabwareId: Selector<DrillDownLabwareId> = createSelector(
-  rootSelector,
-  rootState => rootState.drillDownLabwareId
-)
-
-type ContainersBySlot = { [DeckSlot]: {...Labware, containerId: string} }
-
-const containersBySlot: Selector<ContainersBySlot> = createSelector(
-  getLabwareById,
-  containers => reduce(
-    containers,
-    (acc: ContainersBySlot, containerObj: Labware, containerId: string) => ({
-      ...acc,
-      // NOTE: containerId added in so you still have a reference
-      [containerObj.slot]: {...containerObj, containerId},
-    }),
-    {})
-)
-
-// TODO Ian 2018-07-06 consolidate into types.js
-type IngredGroupFields = {
-  [ingredGroupId: string]: {
-    groupId: string,
-    ...$Exact<IngredInputs>,
-  },
-}
-const allIngredientGroupFields: Selector<AllIngredGroupFields> = createSelector(
-  getLiquidGroupsById,
-  (ingreds) => reduce(
-    ingreds,
-    (acc: IngredGroupFields, ingredGroup: IngredGroupFields, ingredGroupId: string) => ({
-      ...acc,
-      [ingredGroupId]: ingredGroup,
-    }), {})
-)
-
-const allIngredientNamesIds: BaseState => OrderedLiquids = createSelector(
-  getLiquidGroupsById,
-  ingreds => Object.keys(ingreds).map(ingredId =>
-    ({ingredientId: ingredId, name: ingreds[ingredId].name}))
-)
-
-const getLabwareSelectionMode: Selector<boolean> = createSelector(
-  rootSelector,
-  (rootState) => {
-    return rootState.modeLabwareSelection !== false
-  }
-)
-
-const getLiquidGroupsOnDeck: Selector<Array<string>> = createSelector(
-  getLiquidsByLabwareId,
-  (ingredLocationsByLabware) => {
-    let liquidGroups: Set<string> = new Set()
-    forEach(ingredLocationsByLabware, (byWell: $Values<typeof ingredLocationsByLabware>) =>
-      forEach(byWell, (groupContents: $Values<typeof byWell>) => {
-        forEach(groupContents, (contents: $Values<typeof groupContents>, groupId: $Keys<typeof groupContents>) => {
-          if (contents.volume > 0) {
-            liquidGroups.add(groupId)
-          }
-        })
-      })
-    )
-    return [...liquidGroups]
-  }
-)
-
-const getDeckHasLiquid: Selector<boolean> = createSelector(
-  getLiquidGroupsOnDeck,
-  (liquidGroups) => liquidGroups.length > 0
-)
-
-// TODO: prune selectors
-export const selectors = {
-  rootSelector,
-
-  getLiquidGroupsById,
-  getLiquidsByLabwareId,
-  getLiquidNamesById,
-  getLabwareById,
-  getLabwareNicknamesById,
-  getLabwareSelectionMode,
-  getLabwareTypes,
-  getLiquidSelectionOptions,
-  getLiquidGroupsOnDeck,
-  getNextLiquidGroupId,
-  getSavedLabware,
-  getSelectedLabware,
-  getSelectedLabwareId,
-  getSelectedLiquidGroupState,
-  getDrillDownLabwareId,
-
-  allIngredientGroupFields,
-  allIngredientNamesIds,
-  loadedContainersBySlot,
-  containersBySlot,
-  selectedAddLabwareSlot,
-  disposalLabwareOptions,
-  labwareOptions,
-  getDeckHasLiquid,
-}
 
 export default rootReducer
